mock data added

+++ bamazon_db.sql
@@ -6,7 +6,7 @@ CREATE TABLE products (
 item_id INTEGER NOT NULL AUTO_INCREMENT,
 product_name VARCHAR(200) NOT NULL,
 department_name VARCHAR(200) NOT NULL,
-price INTEGER(12),
+price INTEGER(30),
 stock_quantity INTEGER(12) NOT NULL,
 PRIMARY KEY (item_id)
 );
